+++ zwatershed/zwatershed_main.cpp
@@ -42,7 +42,7 @@ double LOW=  .0001;
 double HIGH= .9999;
 bool RECREATE_RG = true;
 
-std::map<std::string,std::list<float>> zwshed_initial_c(const int dimX, const int dimY, const int dimZ, float* affs)
+std::map<std::string,std::list<float>> zwshed_initial_c(const size_t dimX, const size_t dimY, const size_t dimZ, float* affs)
 {
     std::cout << "calculating basic watershed..." << std::endl;
 
@@ -52,7 +52,7 @@ std::map<std::string,std::list<float>> zwshed_initial_c(const int dimX, const in
     affinity_graph_ptr<float> aff(new affinity_graph<float>
                               (boost::extents[dimX][dimY][dimZ][3],
                                boost::fortran_storage_order()));
-    for(int i=0;i<dimX*dimY*dimZ*3;i++)
+    for(size_t i=0;i<dimX*dimY*dimZ*3;i++)
         aff->data()[i] = affs[i];
     std::tie(seg_ref , counts_ref) = watershed<uint64_t>(aff, LOW, HIGH);
 
@@ -71,7 +71,7 @@ std::map<std::string,std::list<float>> zwshed_initial_c(const int dimX, const in
     }
     std::list<float> seg_data = * (new std::list<float>());
     std::list<float> counts_data = * (new std::list<float>());
-    for(int i=0;i<dimX*dimY*dimZ;i++)
+    for(size_t i=0;i<dimX*dimY*dimZ;i++)
         seg_data.push_back(seg_ref->data()[i]);
     for (const auto& x:counts_ref)
         counts_data.push_back(x);
@@ -82,7 +82,7 @@ std::map<std::string,std::list<float>> zwshed_initial_c(const int dimX, const in
  }
 
 
-std::map<std::string,std::vector<double>> merge_with_stats(int dimX,int dimY, int dimZ, uint64_t * gt, float * rgn_graph,
+std::map<std::string,std::vector<double>> merge_with_stats(size_t dimX, size_t dimY, size_t dimZ, uint64_t * gt, float * rgn_graph,
 int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int thresh){
 
     //read data
@@ -90,7 +90,7 @@ int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int th
     volume_ptr<uint64_t> seg(new volume<uint64_t> (boost::extents[dimX][dimY][dimZ], boost::fortran_storage_order()));
     std::vector<std::size_t> counts = * new std::vector<std::size_t>();
     region_graph_ptr<uint64_t,float> rg( new region_graph<uint64_t,float> );
-    for(int i=0;i<dimX*dimY*dimZ;i++){
+    for(size_t i=0;i<dimX*dimY*dimZ;i++){
         gt_ptr->data()[i] = gt[i];
         seg->data()[i] = seg_in[i];
     }
@@ -110,7 +110,7 @@ int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int th
     std::vector<double> r;
     std::vector<double> rg_data; // = * (new std::list<float>());
     std::vector<double> counts_data; // = * (new std::list<float>());
-    for(int i=0;i<dimX*dimY*dimZ;i++)
+    for(size_t i=0;i<dimX*dimY*dimZ;i++)
         seg_vector.push_back(((double)(seg->data()[i])));
     auto x = compare_volumes(*gt_ptr, *seg, dimX,dimY,dimZ);
     r.push_back(std::get<0>(x));
@@ -131,7 +131,7 @@ int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int th
     return returnMap;
 }
 
-std::map<std::string,std::vector<double>> merge_no_stats(int dimX, int dimY, int dimZ, float * rgn_graph,
+std::map<std::string,std::vector<double>> merge_no_stats(size_t dimX, size_t dimY, size_t dimZ, float * rgn_graph,
                                         int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int thresh){
     std::cout << "evaluating..." << std::endl;
 
@@ -139,7 +139,7 @@ std::map<std::string,std::vector<double>> merge_no_stats(int dimX, int dimY, int
     volume_ptr<uint64_t> seg(new volume<uint64_t> (boost::extents[dimX][dimY][dimZ], boost::fortran_storage_order()));
     std::vector<std::size_t> counts = * new std::vector<std::size_t>();
     region_graph_ptr<uint64_t,float> rg( new region_graph<uint64_t,float> );
-    for(int i=0;i<dimX*dimY*dimZ;i++)
+    for(size_t i=0;i<dimX*dimY*dimZ;i++)
         seg->data()[i] = seg_in[i];
     for(int i=0;i<counts_len;i++)
         counts.push_back(counts_in[i]);
@@ -156,7 +156,7 @@ std::map<std::string,std::vector<double>> merge_no_stats(int dimX, int dimY, int
     std::vector<double> seg_vector;
     std::vector<double> rg_data; // = * (new std::list<float>());
     std::vector<double> counts_data; // = * (new std::list<float>());
-    for(int i=0;i<dimX*dimY*dimZ;i++)
+    for(size_t i=0;i<dimX*dimY*dimZ;i++)
         seg_vector.push_back(((double)(seg->data()[i])));
     for ( const auto& e: *rg ){
         rg_data.push_back(std::get<1>(e));
@@ -176,7 +176,7 @@ std::map<std::string,std::vector<double>> merge_no_stats(int dimX, int dimY, int
 
 
 
-std::map<std::string,std::list<float>> zwshed_initial_c_arb(const int dimX, const int dimY, const int dimZ, const uint64_t*node1,
+std::map<std::string,std::list<float>> zwshed_initial_c_arb(const size_t dimX, const size_t dimY, const size_t dimZ, const uint64_t*node1,
                                                const uint64_t*node2, const float*edgeWeight, const int n_edge){
     // read data
     std::cout << "calculating basic watershed..." << std::endl;
@@ -200,7 +200,7 @@ std::map<std::string,std::list<float>> zwshed_initial_c_arb(const int dimX, cons
     }
     std::list<float> seg_data = * (new std::list<float>());
     std::list<float> counts_data = * (new std::list<float>());
-    for(int i=0;i<dimX*dimY*dimZ;i++)
+    for(size_t i=0;i<dimX*dimY*dimZ;i++)
         seg_data.push_back(seg_ref->data()[i]);
     for (const auto& x:counts_ref)
         counts_data.push_back(x);
@@ -213,7 +213,7 @@ std::map<std::string,std::list<float>> zwshed_initial_c_arb(const int dimX, cons
  }
 
 
-std::map<std::string,std::vector<double>> merge_with_stats_arb(int dimX,int dimY, int dimZ, uint64_t * gt, float * rgn_graph,
+std::map<std::string,std::vector<double>> merge_with_stats_arb(size_t dimX,size_t dimY, size_t dimZ, uint64_t * gt, float * rgn_graph,
 int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int thresh){
 
     //read data
@@ -221,7 +221,7 @@ int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int th
     volume_ptr<uint64_t> seg(new volume<uint64_t> (boost::extents[dimX][dimY][dimZ], boost::c_storage_order()));
     std::vector<std::size_t> counts = * new std::vector<std::size_t>();
     region_graph_ptr<uint64_t,float> rg( new region_graph<uint64_t,float> );
-    for(int i=0;i<dimX*dimY*dimZ;i++){
+    for(size_t i=0;i<dimX*dimY*dimZ;i++){
         gt_ptr->data()[i] = gt[i];
         seg->data()[i] = seg_in[i];
     }
@@ -241,7 +241,7 @@ int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int th
     std::vector<double> r;
     std::vector<double> rg_data; // = * (new std::list<float>());
     std::vector<double> counts_data; // = * (new std::list<float>());
-    for(int i=0;i<dimX*dimY*dimZ;i++)
+    for(size_t i=0;i<dimX*dimY*dimZ;i++)
         seg_vector.push_back(((double)(seg->data()[i])));
     auto x = compare_volumes_arb(*gt_ptr, *seg, dimX,dimY,dimZ);
     r.push_back(x.first);
@@ -261,14 +261,14 @@ int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int th
     return returnMap;
 }
 
-std::map<std::string,std::vector<double>> merge_no_stats_arb(int dimX,int dimY, int dimZ, float * rgn_graph,
+std::map<std::string,std::vector<double>> merge_no_stats_arb(size_t dimX,size_t dimY, size_t dimZ, float * rgn_graph,
 int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int thresh){
 
     //read data
     volume_ptr<uint64_t> seg(new volume<uint64_t> (boost::extents[dimX][dimY][dimZ]));
     std::vector<std::size_t> counts = * new std::vector<std::size_t>();
     region_graph_ptr<uint64_t,float> rg( new region_graph<uint64_t,float> );
-    for(int i=0;i<dimX*dimY*dimZ;i++){
+    for(size_t i=0;i<dimX*dimY*dimZ;i++){
         seg->data()[i] = seg_in[i];
     }
     for(int i=0;i<counts_len;i++)
@@ -286,7 +286,7 @@ int rgn_graph_len, uint64_t * seg_in, uint64_t*counts_in, int counts_len, int th
     std::vector<double> seg_vector;
     std::vector<double> rg_data; // = * (new std::list<float>());
     std::vector<double> counts_data; // = * (new std::list<float>());
-    for(int i=0;i<dimX*dimY*dimZ;i++)
+    for(size_t i=0;i<dimX*dimY*dimZ;i++)
         seg_vector.push_back(((double)(seg->data()[i])));
 
     for ( const auto& e: *rg ){
